web/dreamadreamApi/requirements.txt
Django==1.8
django-braces==1.8.1
django-cors-headers==1.1.0
django-debug-toolbar==1.4
django-loaderio==0.3.0
django-oauth-toolkit==0.9.0
djangorestframework==3.1.3
httplib2==0.9.2
oauthlib==1.0.3
PyMySQL==0.6.6
python-dateutil==2.4.2
PyYAML==3.11
requests==2.8.0
sqlparse==0.1.18
tornado==4.3
virtualenv==13.1.2
webapp2==2.5.2